Create play and pause button. Delete player_window.py. All script in one file.
