+++ .idea/dictionaries/igor.xml
@@ -1,6 +1,7 @@
 <component name="ProjectDictionaryState">
   <dictionary name="igor">
     <words>
+      <w>terapsi</w>
       <w>terpsi</w>
     </words>
   </dictionary>
